[메카인] #MMOFighter, Log.h/.cpp 수정 및 Session.action변수 타입 수정

+++ Practice/MMOFighterServerChoi/MMOFighterServerChoi/Session.h
@@ -17,7 +17,7 @@ struct Session
     RingBuffer recvQ;
     RingBuffer sendQ;
 
-    DWORD action = 0;
+    BYTE action = 0;
     BYTE direction = 0;
     short x = 0;
     short y = 0;
